only cast the HourOfDay using unsigned short

+++ src/EnergyPlus/WeatherManager.cc
@@ -2487,8 +2487,8 @@ namespace WeatherManager {
 
         UpdateScheduleValues();
         
-		std::sprintf(time_stamp, "%hu/%hu %hu:",(unsigned short) Month, (unsigned short)DayOfMonth, (unsigned short)(HourOfDay - 1));
-	
+		std::sprintf(time_stamp, "%2d/%2d %hu:", Month, DayOfMonth, (unsigned short)(HourOfDay - 1));
+        	
         CurMnDyHr = time_stamp;
         std::sprintf(day_stamp, "%02d/%02d", Month, DayOfMonth);
         CurMnDy = day_stamp;
